refactor: payment 분리
