+++ src/basic/components/index.tsx
@@ -3,3 +3,5 @@ export * from './Notification';
 export * from './main/Coupon';
 export * from './main/Product';
 export * from './main/Cart';
+export * from './main/Payment';
+export * from './admin/AdminPage';
