+++ src/basic/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useNotification';
 export * from './useProduct';
 export * from './useCart';
 export * from './useCoupon';
+export * from './usePayment';
