+++ src/basic/hooks/usePayment.ts
@@ -1,0 +1,22 @@
+import { useCallback } from 'react';
+import { CartItemType } from '../../types';
+import { CouponType } from '../../types';
+
+export const usePayment = (
+  addNotification: (message: string, type: 'success' | 'error') => void,
+  setCart: (cart: CartItemType[]) => void,
+  setSelectedCoupon: (coupon: CouponType | null) => void,
+  calculateCartTotal: () => { totalBeforeDiscount: number; totalAfterDiscount: number }
+) => {
+  // 주문 완료 처리
+  const completeOrder = useCallback(() => {
+    const orderNumber = `ORD-${Date.now()}`;
+    addNotification(`주문이 완료되었습니다. 주문번호: ${orderNumber}`, 'success');
+    setCart([]);
+    setSelectedCoupon(null);
+  }, [addNotification]);
+
+  // 장바구니 총액 계산
+  const totals = calculateCartTotal();
+  return { totals, completeOrder };
+};
